Major refactoring to add a node abstraction.
The changes allow the simulator to have different
node types defined in node.h.

Now, every node is identified by an universal uuid
rather than by dpids.

In order for the control plane to use the dpid,
there is need to have a mapping between the dpids,
which will be used to create policies, and
the uuid of the corresponding node in the topology.

Another relevant change of this commit is to add
a forward declaration of topology and datapath.
That way, it should be easier to create the Python
binding that will ease the creation of topologies.